running mr on gwas

+++ 02_meta_analysis_17/run_metal.sh
@@ -1,14 +1,8 @@
 #!/bin/bash
 
-#PBS -N meta_17
-#PBS -o job_reports/meta_17-output
-#PBS -e job_reports/meta_17-error
-#PBS -t 1-300
-#PBS -l walltime=12:00:00
-#PBS -l nodes=1:ppn=2
-#PBS -S /bin/bash
-
-
+#SBATCH --job-name=meta17
+#SBATCH --array=1-300
+#SBATCH --nodes=1 --mem=25G --time=0-12:00:00
 
 set -e
 
@@ -17,18 +11,27 @@ set -e
 
 
 echo "Running on ${HOSTNAME}"
+module add R/3.2.3-foss-2016a
+
 start_time=`date +%s`
 
 
+# if [ -n "${1}" ]; then
+#   echo "${1}"
+#   PBS_ARRAYID=${1}
+# fi
+
+# i=${PBS_ARRAYID}
+
 if [ -n "${1}" ]; then
   echo "${1}"
-  PBS_ARRAYID=${1}
+  SLURM_ARRAY_TASK_ID=${1}
 fi
 
-i=${PBS_ARRAYID}
+i=${SLURM_ARRAY_TASK_ID}
 
 
-cd /panfs/panasas01/shared-godmc/godmc_phase2_analysis/02_meta_analysis_17
+cd /mnt/storage/private/mrcieu/research/UKBIOBANK_Phenotypes_App_15825/scripts/godmc_phase2_analysis/02_meta_analysis_17
 
 cohort_dir="../data/17/"
 metal_dir="../scratch/17_${i}"
@@ -59,21 +62,24 @@ echo "" >> ${metal_dir}/${metal_in}
 
 for cohort in ${cohort_dir}*17.tar
 do
-
 	echo $cohort
 	cohortname=$(basename "$cohort" .tar)
+	echo ${cohortname}
 
+	if [ ! -e "${metal_dir}/${cohortname}_${i}.out.gz" ]
+	then
+		mkdir -p ${cohortname}_${i}
+		cd ${cohortname}_${i}
+		tar xvf ../${cohort} results/17/results_${i}.txt.gz
+		cd ../
+		mv ${cohortname}_${i}/results/17/results_${i}.txt.gz ${metal_dir}/${cohortname}_${i}.gz
+		rm -r ${cohortname}_${i}
+		Rscript make_gwama.r ${metal_dir}/${cohortname}_${i}.gz ${metal_dir}/${cohortname}_${i}.out.gz
+		rm -f ${metal_dir}/${cohortname}_${i}.gz
+	else
+		echo "Already produced"
+	fi
 
-	mkdir -p ${cohortname}_${i}
-	cd ${cohortname}_${i}
-	tar xvf ../${cohort} results/17/results_${i}.txt.gz
-	cd ../
-	mv ${cohortname}_${i}/results/17/results_${i}.txt.gz ${metal_dir}/${cohortname}_${i}.gz
-	rm -r ${cohortname}_${i}
-
-
-	Rscript make_gwama.r ${metal_dir}/${cohortname}_${i}.gz ${metal_dir}/${cohortname}_${i}.out.gz
-	rm ${metal_dir}/${cohortname}_${i}.gz
 	echo "PROCESS ${cohortname}_${i}.out.gz" >> ${metal_dir}/${metal_in}
 
 done
@@ -91,7 +97,7 @@ mv 17_${i}1.txt 17_${i}.txt
 gzip 17_${i}.txt
 cd -
 mv ${metal_dir}/17_${i}.txt.* ${result_dir}
-rm -r ${metal_dir}
+# rm -r ${metal_dir}
 
 # GWAMA --filelist ${metal_file} --quantitative
